Remove ipython_genutils
It should not be necessary and has been deprecated fro 7 years. 
There should not be any dependency that uses it, if so then it should be fixed upstream. 

Keeping ipython_genutils  this is problematic as it depends on nose, which is not packaged in many distributions anymore; and we receive bug report.